configs/legoev3: move Pybricks .elf load address
Move the Pybricks .elf load address to 0xC1090000.

The bootlef command copies the sections from the .elf file to where
they are expected to be. Previously the load address was 0xC0007000,
which caused the sections to be copied of the .elf file. This was
probably OK since it lined up just right so that the .text section
was already at the address it was being copied to. But to avoid any
potiential issues, it will be better to load the .elf file somewhere
else where it won't get written over. 0xC1090000 is 1 MiB above where
u-boot.bin is loaded, so it should be safe.