configs/legoev3_defconfig: drop LOCALVERSION from defconfig
Drop LOCALVERSION from the defconfig file. This way local builds won't
use the same LOCALVERSION as the official builds, and we don't have to
remember to bump it when we release a new version. And we don't have
to deal with version commits when we rebase on new upstream commits.